layouts refactorizados
