+++ config/index.ts
@@ -1,1 +1,3 @@
-export { SessionProvider, useSession, getSession } from 'next-auth/react'
+export * from 'next-auth/react'
+export const API = process.env.API_REST
+export const URL_BASE = process.env.NEXTAUTH_URL
