--- consts/config.ts
@@ -1,2 +1,0 @@
-export const API = process.env.API_REST
-export const URL_BASE = process.env.NEXTAUTH_URL
