+++ services/auth/index.ts
@@ -1,2 +1,4 @@
 export * from './loginUser.service'
 export * from './loginProveedor.service'
+export * from './logoutProveedor.service'
+export * from './logoutUser.service'
