+++ utils/fetcher.utility.ts
@@ -1,5 +1,5 @@
 import axios from 'axios'
-import { API } from '../consts/config'
+import { API } from '../config'
 
 export const senergy = axios.create({
   baseURL: API
